function that prints all the elements of a list_t list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,14 +11,14 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int nodes = 0;
+	size_t nodes = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
 			printf("[0] (nil)\n");
 		else
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		h = h->next;
 		nodes++;
 	}
